When creating files in a test scenario, we can define its content via
the `content` attribute. If it's not present, the content will default
to "whatever".
However, this attribute was not taken into account in the capture's
init phase, resulting in inadequate captures and thus test failures or
incorrectness.

+++ dev/capture/local.js
@@ -54,7 +54,7 @@ const setupInitialState = (scenario /*: Scenario */) => {
   if (scenario.init == null) return
   debug('[init]')
   return Promise.each(scenario.init, opts => {
-    let { ino, path: relpath } = opts
+    let { ino, path: relpath, content } = opts
     if (relpath.endsWith('/')) {
       debug('- mkdir', relpath)
       return fse
@@ -66,7 +66,7 @@ const setupInitialState = (scenario /*: Scenario */) => {
     } else {
       debug('- >', relpath)
       return fse
-        .outputFile(abspath(relpath), 'whatever')
+        .outputFile(abspath(relpath), content || 'whatever')
         .then(() => fse.stat(abspath(relpath)))
         .then(stats => {
           mapInode[stats.ino] = ino
@@ -196,6 +196,13 @@ const runAndRecordFSEvents =
     : runAndRecordChokidarEvents
 
 const captureScenario = (scenario /*: Scenario & {path: string} */) => {
+  if (
+    (scenario.platforms && !scenario.platforms.includes(process.platform)) ||
+    (scenario.side && scenario.side !== 'local')
+  ) {
+    return
+  }
+
   return fse
     .emptyDir(syncPath)
     .then(() => fse.emptyDir(outsidePath))
